requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
numpy==1.24.4
tensorflow==2.14.0
pillow==9.4.0
pyjwt==2.8.0
